+++ CHANGELOG.md
@@ -3,6 +3,18 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.28.1] - 2022-07-21
+### Added
+- Implemented some suggestions by `Washai#0161`
+  - Required Ingredients Effected by "Subtract Amount From Inventory" Setting
+    - Required Ingredient amount is now effected by the "Subtract Amount From Inventory"
+    - An item with a reduced amount from inventory will have their text color be red and an inventory crate icon will be displayed on the top left of the item
+  - `+5` and `+1` Multiplier for the Inline Inventory Editor Depending on Event Drops
+    - The `+5` and `+1` buttons will now be modified depending on your event drop buff setting
+    - Quests of different difficulties are not effected by a different difficulty's drop buff setting
+      - `Normal 3x` Drop Buff will not make Hard quests have `+15` and `+3` buttons in the inline editor
+  - Made some changes to `region-update.js` where characters will be sorted by their `unit_name` instead of `unit_id`
+
 ## [2.28.0] - 2022-07-15
 Quest Update (2022-07-15)
 ### Added